Added Player Control

+++ TetriClimber/TetriClimber/Program.cs
@@ -39,7 +39,7 @@ namespace TetriClimber
             get
             {
                 //return ((Form)Form.FromHandle(Window.Handle)).DesktopBounds.Size;
-                return new Size(Constants.Measures.portraitWidth, Constants.Measures.portraitHeight);
+                return new Size((int)Constants.Measures.portraitWidth, (int)Constants.Measures.portraitHeight);
             }
         }
 
@@ -119,8 +119,8 @@ namespace TetriClimber
             int height = (InteractiveSurface.PrimarySurfaceDevice != null)
                             ? InteractiveSurface.PrimarySurfaceDevice.WorkingAreaHeight
                             : Screen.PrimaryScreen.WorkingArea.Height;
-            height = Constants.Measures.portraitHeight; //889;
-            width = Constants.Measures.portraitWidth; //500;
+            height = (int)Constants.Measures.portraitHeight; //889;
+            width = (int)Constants.Measures.portraitWidth; //500;
             Form form = (Form)Form.FromHandle(Window.Handle);
             form.ClientSize = new Size(width, height);
             form.WindowState = (SurfaceEnvironment.IsSurfaceEnvironmentAvailable)
